Make url route equatable
